Update Homework13.java
Adding of Homework - Maps
